Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main(){
-    int r,area;
+    float r,area;
     scanf("%f",&r);
     area = r*3.14;
     print("Area: %.2f",area);
